% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ssd.R
\name{ComputeSSD}
\alias{ComputeSSD}
\title{compute steady-state distribution from transition matrix}
\usage{
ComputeSSD(transition.mat, iter.epsilon = 0.00000001, iter.max = 100000)
}
\arguments{
\item{transition.mat}{Matrix, transition matrix.}

\item{iter.epsilon}{Float, the tolerance of iterative computation of SSD}

\item{iter.max}{Int, the max number of iteration}
}
\value{
Named vector, steady-state distribution.
}
\description{
compute steady-state distribution from transition matrix
}

