% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_transition_mat.R
\name{EstimateTransitionMat}
\alias{EstimateTransitionMat}
\title{estimate transition matrix from distance matrix}
\usage{
EstimateTransitionMat(ice.matrix, distance.matrix, method = "binomial",
  t = 0.001, distance_scale = 1)
}
\arguments{
\item{ice.matrix}{Matrix ICE normalized matrix.}

\item{distance.matrix}{Matrix, shortest-path corrected matrix.}

\item{method, }{String c('binomial', 'Wiener'), binomial or
Wiener Process method to estimate transition matrix,
in the previous manuscript we use binomial method.}
}
\value{
Matrix, transition matrix.
}
\description{
estimate transition matrix from distance matrix
}

