% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ICENormalization}
\alias{ICENormalization}
\title{ICE normalization of Hi-C matrix}
\usage{
ICENormalization(htc.obj, max.iter = 200)
}
\arguments{
\item{htc.obj}{HTC object.}

\item{max.iter}{Integer, maximum number of iteration}
}
\value{
HTC object of the ICE normalized matrix
}
\description{
ICE normalization of Hi-C matrix
}

