% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{LoadRawMatrix}
\alias{LoadRawMatrix}
\title{Load Hi-C raw matrix stored in sparse format and transform it to full matrix}
\usage{
LoadRawMatrix(raw.matrix.path, chr.size, resolution)
}
\arguments{
\item{raw.matrix.path}{String, file path for sparse matrix, contain Hi-C
data for one chromosome.}

\item{chr.size}{Integer, the length of chromosome.}

\item{resolution}{Integer, Hi-C resolution.}
}
\value{
Matrix, Hi-C raw matrix
}
\description{
Load Hi-C raw matrix stored in sparse format and transform it to full matrix
}

