% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov3d.R
\docType{package}
\name{markov3d}
\alias{markov3d}
\alias{markov3d-package}
\title{markov3d: A package to use Markov process to model the dynamic interactions
between proteins and genome and compute the steady-state distribution of
proteins on the genome.}
\description{
The markov3d package provide functions to load data, preprocess data and
compute steady-state distribution. See each function's documatation for help.
}

