% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{MatrixToHTCobj}
\alias{MatrixToHTCobj}
\title{transform Hi-C matrix to HTC object}
\usage{
MatrixToHTCobj(contact.matrix, chr, chr.size, resolution)
}
\arguments{
\item{contact.matrix}{Matrix.}

\item{chr}{Character, the chromosome of the matrix.}

\item{chr.size}{Integer, the length of chromosome.}

\item{resolution}{Integer, Hi-C resolution.}
}
\value{
HTC object of the contact matrix
}
\description{
transform Hi-C matrix to HTC object
}

