% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{RemoveRegions}
\alias{RemoveRegions}
\title{Remove regions (regions without reads mapped, centromere,
regions near centromere)}
\usage{
RemoveRegions(hic.matrix, centromere.region, unmapped.region, resolution,
  n = 10)
}
\arguments{
\item{hic.matrix}{Matrix, contact matrix.}

\item{centromere.region}{Numeric vector with length 2, the first element
should be the start of centromere and the second should be the end.}

\item{unmapped.region}{Numeric vector, regions without mapped reads.}

\item{resolution}{Integer, resolution of Hi-C.}

\item{n}{Integer, the number of bins to be removed around centromere.}
}
\value{
Matrix, Hi-C matrix without centromere and unmapped bins
}
\description{
Remove regions (regions without reads mapped, centromere,
regions near centromere)
}

