% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ShortestPathCorrection}
\alias{ShortestPathCorrection}
\title{shortest-path correction of contact matrix}
\usage{
ShortestPathCorrection(htc.obj, resolution = FALSE)
}
\arguments{
\item{htc.obj}{HTC object.}
}
\value{
HTC object of the shortest-path corrected matrix.
}
\description{
shortest-path correction of contact matrix
}

